Token-based Authentication

+++ app/Http/Controllers/MeetingController.php
@@ -6,14 +6,18 @@ use Carbon\Carbon;
 use Illuminate\Http\Request;
 
 use App\Http\Requests;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 use App\Meeting;
+use JWTAuth;
 
 class MeetingController extends Controller
 {
     public function __construct()
     {
-        // $this->middleware('name');
+        $this->middleware('jwt.auth', ['only' => [
+            'store', 'update', 'destroy'
+        ]]);
     }
 
     /**
@@ -49,14 +53,18 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+
+        if(!$user = JWTAuth::parseToken()->authenticate())
+        {
+            return response()->json(['msg' => "User not found", 404]);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = new Meeting([
             'time' => Carbon::createFromFormat('YmdHie', $time),
@@ -91,7 +99,12 @@ class MeetingController extends Controller
      */
     public function show($id)
     {
-        $meeting = Meeting::with('users')->where('id', $id)->firstOrFail();
+        try {
+            $meeting = Meeting::with('users')->where('id', $id)->firstOrFail();
+        } catch (ModelNotFoundException $e) {
+            return response()->json(['msg' => 'Could not find meeting with id = '.$id], 500);
+        }
+
         $meeting->view_meetings = [
             'href' => 'api/v1/meeting',
             'method' => 'GET'
@@ -116,14 +129,18 @@ class MeetingController extends Controller
         $this->validate($request, [
             'title' => 'required',
             'description' => 'required',
-            'time' => 'required|date_format:YmdHie',
-            'user_id' => 'required'
+            'time' => 'required|date_format:YmdHie'
         ]);
+        
+        if(!$user = JWTAuth::parseToken()->authenticate())
+        {
+            return response()->json(['msg' => "User not found", 404]);
+        }
 
         $title = $request->input('title');
         $description = $request->input('description');
         $time = $request->input('time');
-        $user_id = $request->input('user_id');
+        $user_id = $user->id;
 
         $meeting = Meeting::with('users')->findOrFail($id);
         //return response()->json($request->input('user_id'), 200);
@@ -157,8 +174,22 @@ class MeetingController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
-        $meeting = Meeting::findOrFail($id);
+    {      
+        try {
+                $meeting = Meeting::findOrFail($id);
+        } catch (ModelNotFoundException $e) {
+            return response()->json(['msg' => 'Could not find meeting with id = '.$id], 500);
+        }
+
+        if(!$user = JWTAuth::parseToken()->authenticate())
+        {
+            return response()->json(['msg' => "User not found", 404]);
+        }
+
+        if (!$meeting->users()->where('users.id', $user->id)->first()) {
+            return response()->json(['msg' => 'user not registered for meeting, update not successful'], 401);
+        };
+
         $users = $meeting->users;
         $meeting->users()->detach();
         if (!$meeting->delete()) {
